splitting fatigue of ding model
