+++ cocofest/models/ding2003.py
@@ -13,11 +13,13 @@ from bioptim import (
 
 class DingModelFrequency:
     """
-    This is a custom models that inherits from bioptim. CustomModel.
-    As CustomModel is an abstract class, some methods are mandatory and must be implemented.
-    Such as serialize, name_dof, nb_state and name.
+    This is a custom model of the Bioptim package. As CustomModel, some methods are mandatory and must be implemented.
+    to make it work with bioptim.
 
-    This is the Ding 2003 model using the stimulation frequency in input.
+    This is the Ding 2003 model using the stimulation frequency as a control input.
+
+    Notes
+    -----
 
     Ding, J., Wexler, A. S., & Binder-Macleod, S. A. (2003).
     Mathematical models for fatigue minimization during functional electrical stimulation.
@@ -27,11 +29,9 @@ class DingModelFrequency:
     def __init__(
         self,
         name: str = None,
-        with_fatigue: bool = True,
         sum_stim_truncation: int = None,
     ):
         self._name = name
-        self._with_fatigue = with_fatigue
         self._sum_stim_truncation = sum_stim_truncation
         # ---- Custom values for the example ---- #
         self.tauc = 0.020  # Value from Ding's experimentation [1] (s)
@@ -42,11 +42,6 @@ class DingModelFrequency:
         self.tau1_rest = 0.050957  # Value from Ding's experimentation [1] (s)
         self.tau2 = 0.060  # Close value from Ding's experimentation [2] (s)
         self.km_rest = 0.103  # Value from Ding's experimentation [1] (unitless)
-        # ---- Fatigue models ---- #
-        self.alpha_a = -4.0 * 10e-7  # Value from Ding's experimentation [1] (s^-2)
-        self.alpha_tau1 = 2.1 * 10e-5  # Value from Ding's experimentation [1] (N^-1)
-        self.tau_fat = 127  # Value from Ding's experimentation [1] (s)
-        self.alpha_km = 1.9 * 10e-8  # Value from Ding's experimentation [1] (s^-1.N^-1)
 
     def set_a_rest(self, model, a_rest: MX | float):
         # models is required for bioptim compatibility
@@ -61,51 +56,19 @@ class DingModelFrequency:
     def set_tau2(self, model, tau2: MX | float):
         self.tau2 = tau2
 
-    def set_alpha_a(self, model, alpha_a: MX | float):
-        self.alpha_a = alpha_a
-
-    def set_alpha_km(self, model, alpha_km: MX | float):
-        self.alpha_km = alpha_km
-
-    def set_alpha_tau1(self, model, alpha_tau1: MX | float):
-        self.alpha_tau1 = alpha_tau1
-
-    def set_tau_fat(self, model, tau_fat: MX | float):
-        self.tau_fat = tau_fat
-
     def standard_rest_values(self) -> np.array:
         """
         Returns
         -------
-        The rested values of Cn, F, A, Tau1, Km
+        The rested values of the states Cn, F
         """
-        return (
-            np.array([[0], [0], [self.a_rest], [self.tau1_rest], [self.km_rest]])
-            if self._with_fatigue
-            else np.array([[0], [0]])
-        )
+        return np.array([[0], [0]])
 
     # ---- Absolutely needed methods ---- #
     def serialize(self) -> tuple[Callable, dict]:
         # This is where you can serialize your models
         # This is useful if you want to save your models and load it later
         return (
-            (
-                DingModelFrequency,
-                {
-                    "tauc": self.tauc,
-                    "a_rest": self.a_rest,
-                    "tau1_rest": self.tau1_rest,
-                    "km_rest": self.km_rest,
-                    "tau2": self.tau2,
-                    "alpha_a": self.alpha_a,
-                    "alpha_tau1": self.alpha_tau1,
-                    "alpha_km": self.alpha_km,
-                    "tau_fat": self.tau_fat,
-                },
-            )
-            if self._with_fatigue
-            else (
                 DingModelFrequency,
                 {
                     "tauc": self.tauc,
@@ -115,23 +78,22 @@ class DingModelFrequency:
                     "tau2": self.tau2,
                 },
             )
-        )
 
     # ---- Needed for the example ---- #
     @property
     def name_dof(self) -> list[str]:
-        return ["Cn", "F", "A", "Tau1", "Km"] if self._with_fatigue else ["Cn", "F"]
+        return ["Cn", "F"]
 
     @property
     def nb_state(self) -> int:
-        return 5 if self._with_fatigue else 2
+        return 2
 
     @property
     def name(self) -> None | str:
         return self._name
 
     # ---- Model's dynamics ---- #
-    def system_dynamics_without_fatigue(
+    def system_dynamics(
         self,
         cn: MX,
         f: MX,
@@ -160,48 +122,6 @@ class DingModelFrequency:
         cn_dot = self.cn_dot_fun(cn, r0, t, t_stim_prev=t_stim_prev)  # Equation n°1
         f_dot = self.f_dot_fun(cn, f, self.a_rest, self.tau1_rest, self.km_rest)  # Equation n°2
         return vertcat(cn_dot, f_dot)
-
-    def system_dynamics_with_fatigue(
-        self,
-        cn: MX,
-        f: MX,
-        a: MX = None,
-        tau1: MX = None,
-        km: MX = None,
-        t: MX = None,
-        t_stim_prev: list[MX] | list[float] = None,
-    ) -> MX:
-        """
-        The system dynamics is the function that describes the models.
-
-        Parameters
-        ----------
-        cn: MX
-            The value of the ca_troponin_complex (unitless)
-        f: MX
-            The value of the force (N)
-        a: MX
-            The value of the scaling factor (unitless)
-        tau1: MX
-            The value of the time_state_force_no_cross_bridge (ms)
-        km: MX
-            The value of the cross_bridges (unitless)
-        t: MX
-            The current time at which the dynamics is evaluated (ms)
-        t_stim_prev: list[MX]
-            The time list of the previous stimulations (ms)
-
-        Returns
-        -------
-        The value of the derivative of each state dx/dt at the current time t
-        """
-        r0 = km + self.r0_km_relationship  # Simplification
-        cn_dot = self.cn_dot_fun(cn, r0, t, t_stim_prev=t_stim_prev)  # Equation n°1
-        f_dot = self.f_dot_fun(cn, f, a, tau1, km)  # Equation n°2
-        a_dot = self.a_dot_fun(a, f)  # Equation n°5
-        tau1_dot = self.tau1_dot_fun(tau1, f)  # Equation n°9
-        km_dot = self.km_dot_fun(km, f)  # Equation n°11
-        return vertcat(cn_dot, f_dot, a_dot, tau1_dot, km_dot)
 
     def exp_time_fun(self, t: MX, t_stim_i: MX) -> MX | float:
         """
@@ -306,51 +226,6 @@ class DingModelFrequency:
         """
         return a * (cn / (km + cn)) - (f / (tau1 + self.tau2 * (cn / (km + cn))))  # Equation n°2
 
-    def a_dot_fun(self, a: MX, f: MX) -> MX | float:
-        """
-        Parameters
-        ----------
-        a: MX
-            The previous step value of scaling factor (unitless)
-        f: MX
-            The previous step value of force (N)
-
-        Returns
-        -------
-        The value of the derivative scaling factor (unitless)
-        """
-        return -(a - self.a_rest) / self.tau_fat + self.alpha_a * f  # Equation n°5
-
-    def tau1_dot_fun(self, tau1: MX, f: MX) -> MX | float:
-        """
-        Parameters
-        ----------
-        tau1: MX
-            The previous step value of time_state_force_no_cross_bridge (ms)
-        f: MX
-            The previous step value of force (N)
-
-        Returns
-        -------
-        The value of the derivative time_state_force_no_cross_bridge (ms)
-        """
-        return -(tau1 - self.tau1_rest) / self.tau_fat + self.alpha_tau1 * f  # Equation n°9
-
-    def km_dot_fun(self, km: MX, f: MX) -> MX | float:
-        """
-        Parameters
-        ----------
-        km: MX
-            The previous step value of cross_bridges (unitless)
-        f: MX
-            The previous step value of force (N)
-
-        Returns
-        -------
-        The value of the derivative cross_bridges (unitless)
-        """
-        return -(km - self.km_rest) / self.tau_fat + self.alpha_km * f  # Equation n°11
-
     @staticmethod
     def dynamics(
         time: MX,
@@ -385,22 +260,8 @@ class DingModelFrequency:
         The derivative of the states in the tuple[MX] format
         """
 
-        return (
-            DynamicsEvaluation(
-                dxdt=nlp.model.system_dynamics_with_fatigue(
-                    cn=states[0],
-                    f=states[1],
-                    a=states[2],
-                    tau1=states[3],
-                    km=states[4],
-                    t=time,
-                    t_stim_prev=stim_apparition,
-                ),
-                defects=None,
-            )
-            if nlp.model._with_fatigue
-            else DynamicsEvaluation(
-                dxdt=nlp.model.system_dynamics_without_fatigue(
+        return DynamicsEvaluation(
+                dxdt=nlp.model.system_dynamics(
                     cn=states[0],
                     f=states[1],
                     t=time,
@@ -408,7 +269,6 @@ class DingModelFrequency:
                 ),
                 defects=None,
             )
-        )
 
     def declare_ding_variables(self, ocp: OptimalControlProgram, nlp: NonLinearProgram):
         """
@@ -423,10 +283,6 @@ class DingModelFrequency:
         """
         self.configure_ca_troponin_complex(ocp=ocp, nlp=nlp, as_states=True, as_controls=False)
         self.configure_force(ocp=ocp, nlp=nlp, as_states=True, as_controls=False)
-        if self._with_fatigue:
-            self.configure_scaling_factor(ocp=ocp, nlp=nlp, as_states=True, as_controls=False)
-            self.configure_time_state_force_no_cross_bridge(ocp=ocp, nlp=nlp, as_states=True, as_controls=False)
-            self.configure_cross_bridges(ocp=ocp, nlp=nlp, as_states=True, as_controls=False)
         stim_apparition = self.get_stim_prev(ocp, nlp)
         ConfigureProblem.configure_dynamics_function(ocp, nlp, dyn_func=self.dynamics, stim_apparition=stim_apparition)
 
@@ -495,114 +351,6 @@ class DingModelFrequency:
         ConfigureProblem.configure_new_variable(
             name,
             name_f,
-            ocp,
-            nlp,
-            as_states,
-            as_controls,
-            as_states_dot,
-        )
-
-    @staticmethod
-    def configure_scaling_factor(
-        ocp: OptimalControlProgram,
-        nlp: NonLinearProgram,
-        as_states: bool,
-        as_controls: bool,
-        as_states_dot: bool = False,
-    ):
-        """
-        Configure a new variable of the scaling factor (N/ms)
-
-        Parameters
-        ----------
-        ocp: OptimalControlProgram
-            A reference to the ocp
-        nlp: NonLinearProgram
-            A reference to the phase
-        as_states: bool
-            If the generalized coordinates should be a state
-        as_controls: bool
-            If the generalized coordinates should be a control
-        as_states_dot: bool
-            If the generalized velocities should be a state_dot
-        """
-        name = "A"
-        name_a = [name]
-        ConfigureProblem.configure_new_variable(
-            name,
-            name_a,
-            ocp,
-            nlp,
-            as_states,
-            as_controls,
-            as_states_dot,
-        )
-
-    @staticmethod
-    def configure_time_state_force_no_cross_bridge(
-        ocp: OptimalControlProgram,
-        nlp: NonLinearProgram,
-        as_states: bool,
-        as_controls: bool,
-        as_states_dot: bool = False,
-    ):
-        """
-        Configure a new variable for time constant of force decline at the absence of strongly bound cross-bridges (ms)
-
-        Parameters
-        ----------
-        ocp: OptimalControlProgram
-            A reference to the ocp
-        nlp: NonLinearProgram
-            A reference to the phase
-        as_states: bool
-            If the generalized coordinates should be a state
-        as_controls: bool
-            If the generalized coordinates should be a control
-        as_states_dot: bool
-            If the generalized velocities should be a state_dot
-        """
-        name = "Tau1"
-        name_tau1 = [name]
-        ConfigureProblem.configure_new_variable(
-            name,
-            name_tau1,
-            ocp,
-            nlp,
-            as_states,
-            as_controls,
-            as_states_dot,
-        )
-
-    @staticmethod
-    def configure_cross_bridges(
-        ocp: OptimalControlProgram,
-        nlp: NonLinearProgram,
-        as_states: bool,
-        as_controls: bool,
-        as_states_dot: bool = False,
-    ):
-        """
-        Configure a new variable for sensitivity of strongly bound cross-bridges to Cn (unitless)
-
-        Parameters
-        ----------
-        ocp: OptimalControlProgram
-            A reference to the ocp
-        nlp: NonLinearProgram
-            A reference to the phase
-        as_states: bool
-            If the generalized coordinates should be a state
-        as_controls: bool
-            If the generalized coordinates should be a control
-        as_states_dot: bool
-            If the generalized velocities should be a state_dot
-        """
-        name = "Km"
-        name_km = [name]
-        ConfigureProblem.configure_new_variable(
-            name,
-            name_km,
             ocp,
             nlp,
             as_states,
